+++ VsHelix/.vsextension/string-resources.json
@@ -1,0 +1,3 @@
+{
+  "VsHelix.Command1.DisplayName": "Sample Remote Command"
+}
