+++ VsHelix/Command1.cs
@@ -4,7 +4,7 @@ using Microsoft.VisualStudio.Extensibility;
 using Microsoft.VisualStudio.Extensibility.Commands;
 using Microsoft.VisualStudio.Extensibility.Shell;
 
-namespace VxHelix3
+namespace VsHelix
 {
     /// <summary>
     /// Command1 handler.
@@ -26,7 +26,7 @@ namespace VxHelix3
         }
 
         /// <inheritdoc />
-        public override CommandConfiguration CommandConfiguration => new("%VxHelix3.Command1.DisplayName%")
+        public override CommandConfiguration CommandConfiguration => new("%VsHelix.Command1.DisplayName%")
         {
             // Use this object initializer to set optional parameters for the command. The required parameter,
             // displayName, is set above. DisplayName is localized and references an entry in .vsextension\string-resources.json.
