+++ VsHelix/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "VxHelix3": {
+    "VsHelix": {
       "commandName": "Project"
     }
   }
