--- VxHelix3/.vsextension/string-resources.json
@@ -1,3 +1,0 @@
-{
-  "VxHelix3.Command1.DisplayName": "Sample Remote Command"
-}
